requirements.txt
numpy
pyyaml
opencv-python
open3d
matplotlib
